Move net::addr tests to their own files
Avoid recompiling std when tests changed
